requirements.txt
streamlit
pandas
pymysql
python-dotenv
openai
sqlalchemy
httpx
